refactor sync prebuilt analyze tests
