+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_receipt.py
@@ -8,46 +8,14 @@ import pytest
 import functools
 from io import BytesIO
 from datetime import date, time
-from azure.core.exceptions import ClientAuthenticationError, ServiceRequestError, HttpResponseError
-from azure.core.credentials import AzureKeyCredential
-from azure.ai.formrecognizer._generated.v2_1.models import AnalyzeOperationResult
-from azure.ai.formrecognizer._generated.v2021_09_30_preview.models import AnalyzeResultOperation
-from azure.ai.formrecognizer._response_handlers import prepare_prebuilt_models
-from azure.ai.formrecognizer import DocumentAnalysisClient, FormRecognizerClient, FormContentType, FormRecognizerApiVersion, AnalyzeResult
+from azure.ai.formrecognizer import FormRecognizerClient, FormContentType, FormRecognizerApiVersion
 from testcase import FormRecognizerTest
 from preparers import GlobalClientPreparer as _GlobalClientPreparer
 from preparers import FormRecognizerPreparer
 
-GlobalClientPreparer = functools.partial(_GlobalClientPreparer, DocumentAnalysisClient)
 GlobalClientPreparerV2 = functools.partial(_GlobalClientPreparer, FormRecognizerClient)
 
 class TestReceiptFromStream(FormRecognizerTest):
-
-    @FormRecognizerPreparer()
-    def test_receipt_bad_endpoint(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        with open(self.receipt_jpg, "rb") as fd:
-            myfile = fd.read()
-        with self.assertRaises(ServiceRequestError):
-            client = DocumentAnalysisClient("http://notreal.azure.com", AzureKeyCredential(formrecognizer_test_api_key))
-            poller = client.begin_analyze_document("prebuilt-receipt", myfile)
-
-    @FormRecognizerPreparer()
-    def test_authentication_bad_key(self, formrecognizer_test_endpoint, formrecognizer_test_api_key):
-        client = DocumentAnalysisClient(formrecognizer_test_endpoint, AzureKeyCredential("xxxx"))
-        with self.assertRaises(ClientAuthenticationError):
-            poller = client.begin_analyze_document("prebuilt-receipt", b"xx")
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_fail_passing_content_type(self, client):
-        with open(self.receipt_png, "rb") as fd:
-            myfile = fd.read()
-        with self.assertRaises(TypeError):
-            poller = client.begin_analyze_document(
-                "prebuilt-receipt",
-                myfile,
-                content_type=FormContentType.IMAGE_PNG
-            )
 
     @FormRecognizerPreparer()
     @GlobalClientPreparerV2()
@@ -62,31 +30,11 @@ class TestReceiptFromStream(FormRecognizerTest):
         assert result is not None
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_damaged_file_passed_as_bytes(self, client):
-        damaged_pdf = b"\x25\x50\x44\x46\x55\x55\x55"  # still has correct bytes to be recognized as PDF
-        with self.assertRaises(HttpResponseError):
-            poller = client.begin_analyze_document(
-                "prebuilt-receipt",
-                damaged_pdf
-            )
-
-    @FormRecognizerPreparer()
     @GlobalClientPreparerV2()
     def test_damaged_file_bytes_fails_autodetect_content_type(self, client):
         damaged_pdf = b"\x50\x44\x46\x55\x55\x55"  # doesn't match any magic file numbers
         with self.assertRaises(ValueError):
             poller = client.begin_recognize_receipts(
-                damaged_pdf
-            )
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_damaged_file_passed_as_bytes_io(self, client):
-        damaged_pdf = BytesIO(b"\x25\x50\x44\x46\x55\x55\x55")  # still has correct bytes to be recognized as PDF
-        with self.assertRaises(HttpResponseError):
-            poller = client.begin_analyze_document(
-                "prebuilt-receipt",
                 damaged_pdf
             )
 
@@ -99,19 +47,6 @@ class TestReceiptFromStream(FormRecognizerTest):
             poller = client.begin_recognize_receipts(
                 damaged_pdf
             )
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_blank_page(self, client):
-
-        with open(self.blank_pdf, "rb") as fd:
-            blank = fd.read()
-        poller = client.begin_analyze_document(
-            "prebuilt-receipt",
-            blank
-        )
-        result = poller.result()
-        assert result is not None
 
     @FormRecognizerPreparer()
     @GlobalClientPreparerV2()
@@ -132,120 +67,6 @@ class TestReceiptFromStream(FormRecognizerTest):
                 "https://badurl.jpg",
                 content_type="application/json"
             )
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_auto_detect_unsupported_stream_content(self, client):
-
-        with open(self.unsupported_content_py, "rb") as fd:
-            myfile = fd.read()
-
-        with self.assertRaises(HttpResponseError):
-            poller = client.begin_analyze_document(
-                "prebuilt-receipt",
-                myfile
-            )
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_stream_transform_png(self, client):
-        responses = []
-
-        def callback(raw_response, _, headers):
-            analyze_result = client._deserialize(AnalyzeResultOperation, raw_response)
-            extracted_receipt = AnalyzeResult._from_generated(analyze_result.analyze_result)
-            responses.append(analyze_result)
-            responses.append(extracted_receipt)
-
-        with open(self.receipt_png, "rb") as fd:
-            myfile = fd.read()
-
-        poller = client.begin_analyze_document(
-            "prebuilt-receipt",
-            document=myfile,
-            cls=callback
-        )
-
-        result = poller.result()
-        raw_analyze_result = responses[0].analyze_result
-        returned_model = responses[1]
-
-        # Check AnalyzeResult
-        assert returned_model.model_id == raw_analyze_result.model_id
-        assert returned_model.api_version == raw_analyze_result.api_version
-        assert returned_model.content == raw_analyze_result.content
-        
-        self.assertDocumentPagesTransformCorrect(returned_model.pages, raw_analyze_result.pages)
-        self.assertDocumentTransformCorrect(returned_model.documents, raw_analyze_result.documents)
-        self.assertDocumentTablesTransformCorrect(returned_model.tables, raw_analyze_result.tables)
-        self.assertDocumentKeyValuePairsTransformCorrect(returned_model.key_value_pairs, raw_analyze_result.key_value_pairs)
-        self.assertDocumentEntitiesTransformCorrect(returned_model.entities, raw_analyze_result.entities)
-        self.assertDocumentStylesTransformCorrect(returned_model.styles, raw_analyze_result.styles)
-
-        # check page range
-        assert len(raw_analyze_result.pages) == len(returned_model.pages)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_stream_transform_jpg(self, client):
-        responses = []
-
-        def callback(raw_response, _, headers):
-            analyze_result = client._deserialize(AnalyzeResultOperation, raw_response)
-            extracted_receipt = AnalyzeResult._from_generated(analyze_result.analyze_result)
-            responses.append(analyze_result)
-            responses.append(extracted_receipt)
-
-        with open(self.receipt_jpg, "rb") as fd:
-            myfile = fd.read()
-
-        poller = client.begin_analyze_document(
-            "prebuilt-receipt",
-            document=myfile,
-            cls=callback
-        )
-
-        result = poller.result()
-        raw_analyze_result = responses[0].analyze_result
-        returned_model = responses[1]
-
-        # Check AnalyzeResult
-        assert returned_model.model_id == raw_analyze_result.model_id
-        assert returned_model.api_version == raw_analyze_result.api_version
-        assert returned_model.content == raw_analyze_result.content
-        
-        self.assertDocumentPagesTransformCorrect(returned_model.pages, raw_analyze_result.pages)
-        self.assertDocumentTransformCorrect(returned_model.documents, raw_analyze_result.documents)
-        self.assertDocumentTablesTransformCorrect(returned_model.tables, raw_analyze_result.tables)
-        self.assertDocumentKeyValuePairsTransformCorrect(returned_model.key_value_pairs, raw_analyze_result.key_value_pairs)
-        self.assertDocumentEntitiesTransformCorrect(returned_model.entities, raw_analyze_result.entities)
-        self.assertDocumentStylesTransformCorrect(returned_model.styles, raw_analyze_result.styles)
-
-        # check page range
-        assert len(raw_analyze_result.pages) == len(returned_model.pages)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_png(self, client):
-
-        with open(self.receipt_png, "rb") as stream:
-            poller = client.begin_analyze_document("prebuilt-receipt", stream)
-
-        result = poller.result()
-        assert len(result.documents) == 1
-        receipt = result.documents[0]
-        assert receipt.fields.get("MerchantAddress").value, '123 Main Street Redmond ==  WA 98052'
-        assert receipt.fields.get("MerchantName").value ==  'Contoso'
-        assert receipt.fields.get("Subtotal").value ==  1098.99
-        assert receipt.fields.get("Tax").value ==  104.4
-        assert receipt.fields.get("Total").value ==  1203.39
-        assert receipt.fields.get("TransactionDate").value == date(year=2019, month=6, day=10)
-        assert receipt.fields.get("TransactionTime").value == time(hour=13, minute=59, second=0)
-        receipt_type = receipt.fields.get("ReceiptType")
-        assert receipt_type.confidence is not None
-        assert receipt_type.value ==  'Itemized'
-
-        assert len(result.pages) == 1
 
     @FormRecognizerPreparer()
     @GlobalClientPreparerV2()
@@ -281,121 +102,6 @@ class TestReceiptFromStream(FormRecognizerTest):
         assert receipt_type.value ==  'Itemized'
 
     @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_multipage(self, client):
-
-        with open(self.multipage_receipt_pdf, "rb") as fd:
-            receipt = fd.read()
-        poller = client.begin_analyze_document("prebuilt-receipt", receipt)
-        result = poller.result()
-
-        d = result.to_dict()
-        result = AnalyzeResult.from_dict(d)
-
-        assert len(result.documents) == 2
-        receipt = result.documents[0]
-        assert receipt.fields.get("MerchantAddress").value, '123 Main Street Redmond ==  WA 98052'
-        assert receipt.fields.get("MerchantName").value ==  'Contoso'
-        assert receipt.fields.get("MerchantPhoneNumber").value ==  '+19876543210'
-        assert receipt.fields.get("Subtotal").value ==  11.7
-        assert receipt.fields.get("Tax").value ==  1.17
-        assert receipt.fields.get("Tip").value ==  1.63
-        assert receipt.fields.get("Total").value ==  14.5
-        assert receipt.fields.get("TransactionDate").value == date(year=2019, month=6, day=10)
-        assert receipt.fields.get("TransactionTime").value == time(hour=13, minute=59, second=0)
-        receipt_type = receipt.fields.get("ReceiptType")
-        assert receipt_type.confidence is not None
-        assert receipt_type.value ==  'Itemized'
-        receipt = result.documents[1]
-        assert receipt.fields.get("MerchantAddress").value, '123 Main Street Redmond ==  WA 98052'
-        assert receipt.fields.get("MerchantName").value ==  'Contoso'
-        assert receipt.fields.get("Subtotal").value ==  1098.99
-        assert receipt.fields.get("Tax").value ==  104.4
-        assert receipt.fields.get("Total").value ==  1203.39
-        assert receipt.fields.get("TransactionDate").value == date(year=2019, month=6, day=10)
-        assert receipt.fields.get("TransactionTime").value == time(hour=13, minute=59, second=0)
-        receipt_type = receipt.fields.get("ReceiptType")
-        assert receipt_type.confidence is not None
-        assert receipt_type.value ==  'Itemized'
-
-        assert len(result.pages) == 2
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_multipage_transform(self, client):
-
-        responses = []
-
-        def callback(raw_response, _, headers):
-            analyze_result = client._deserialize(AnalyzeResultOperation, raw_response)
-            extracted_receipt = AnalyzeResult._from_generated(analyze_result.analyze_result)
-            responses.append(analyze_result)
-            responses.append(extracted_receipt)
-
-        with open(self.multipage_receipt_pdf, "rb") as fd:
-            myfile = fd.read()
-
-        poller = client.begin_analyze_document(
-            "prebuilt-receipt",
-            document=myfile,
-            cls=callback
-        )
-
-        result = poller.result()
-        raw_analyze_result = responses[0].analyze_result
-        d = responses[1].to_dict()
-        returned_model = AnalyzeResult.from_dict(d)
-
-        # Check AnalyzeResult
-        assert returned_model.model_id == raw_analyze_result.model_id
-        assert returned_model.api_version == raw_analyze_result.api_version
-        assert returned_model.content == raw_analyze_result.content
-
-        self.assertDocumentPagesTransformCorrect(returned_model.pages, raw_analyze_result.pages)
-        self.assertDocumentTransformCorrect(returned_model.documents, raw_analyze_result.documents)
-        self.assertDocumentTablesTransformCorrect(returned_model.tables, raw_analyze_result.tables)
-        self.assertDocumentKeyValuePairsTransformCorrect(returned_model.key_value_pairs, raw_analyze_result.key_value_pairs)
-        self.assertDocumentEntitiesTransformCorrect(returned_model.entities, raw_analyze_result.entities)
-        self.assertDocumentStylesTransformCorrect(returned_model.styles, raw_analyze_result.styles)
-
-        # check page range
-        assert len(raw_analyze_result.pages) == len(returned_model.pages)
-
-    @pytest.mark.live_test_only
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_continuation_token(self, client):
-
-        with open(self.receipt_jpg, "rb") as fd:
-            receipt = fd.read()
-
-        initial_poller = client.begin_analyze_document("prebuilt-receipt", receipt)
-        cont_token = initial_poller.continuation_token()
-        poller = client.begin_analyze_document("prebuilt-receipt", None, continuation_token=cont_token)
-        result = poller.result()
-        assert result is not None
-        initial_poller.wait()  # necessary so azure-devtools doesn't throw assertion error
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_locale_specified(self, client):
-        with open(self.receipt_jpg, "rb") as fd:
-            receipt = fd.read()
-        poller = client.begin_analyze_document("prebuilt-receipt", receipt, locale="en-IN")
-        assert 'en-IN' == poller._polling_method._initial_response.http_response.request.query['locale']
-        result = poller.result()
-        assert result
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_receipt_locale_error(self, client):
-        with open(self.receipt_jpg, "rb") as fd:
-            receipt = fd.read()
-        with pytest.raises(HttpResponseError) as e:
-            client.begin_analyze_document("prebuilt-receipt", receipt, locale="not a locale")
-        assert "InvalidArgument" == e.value.error.code
-
-    @FormRecognizerPreparer()
     @GlobalClientPreparerV2(client_kwargs={"api_version": FormRecognizerApiVersion.V2_0})
     def test_receipt_locale_v2(self, client):
         with open(self.receipt_jpg, "rb") as fd:
@@ -403,14 +109,3 @@ class TestReceiptFromStream(FormRecognizerTest):
         with pytest.raises(ValueError) as e:
             client.begin_recognize_receipts(receipt, locale="en-US")
         assert "'locale' is only available for API version V2_1 and up" in str(e.value)
-
-    @FormRecognizerPreparer()
-    @GlobalClientPreparer()
-    def test_pages_kwarg_specified(self, client):
-        with open(self.receipt_jpg, "rb") as fd:
-            receipt = fd.read()
-
-        poller = client.begin_analyze_document("prebuilt-receipt", receipt, pages="1")
-        assert '1' == poller._polling_method._initial_response.http_response.request.query['pages']
-        result = poller.result()
-        assert result
